requirements.txt
streamlit
numpy
pillow
matplotlib
opencv-python-headless